[roll] Roll infra/recipes [fuchsia_cipd_roller] s/step_summary_text/step_text
I don't fully understand the difference between step_text and
step_summary_text, but it seems that when the `step_summary_text` starts
with a newline, it's dumped into a step log instead of showing up in the
build results page, which is not the intended effect. So just use
`step_text` everywhere in this recipe.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/453968
Original-Revision: 0637088c747e86f881483e040956d4904cdc7752
GitOrigin-RevId: 3b7e994c773ec846b145905e0d4d2047bb95121c
Change-Id: I2b4c481268247ca6df41602a0f3f814085c225ae