[roll] Roll fuchsia [msd-vsi-vip] Kill the context if SubmitCommandBuffer fails.
Currently we have no way of knowing whether a batch was successfully
completed or not, as we just drop it if SubmitCommandBuffer fails (which
triggers the expected semaphore). This should help with root
causing the current flake, where it seems like ResetAfterSubmit
is actually failing but not being caught by the test.

BUG=60383

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453585
Original-Revision: 0e5c4956d9bcc2a04a5c66e61ac5898fd0e0c2b1
GitOrigin-RevId: 039e9de54865ab0049ce18310804ecdd543a604c
Change-Id: I4cc833b86b7b3156f6e630404a56b47b3d547a3a