[roll] Roll fuchsia [wlan][common] Acknowledge US assumption in channel checker
This CL modifies the name of channel checking function "is_valid()" to
include "_in_us", to indicate to users that the function is only
correct for US regulations.

It also makes a small modification to reflect that channels up to 14
are included in the 2GHz band.

Test: fx test wlan-common-tests
Original-Bug: 57664
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453717
Original-Revision: 348c3160dac42e362ed44c3aa4c8dac82b585e93
GitOrigin-RevId: 74ecf3cf6c9e64c1a2149d680df5947bd4e8db05
Change-Id: I6ebf579bd8154773bb4763aa0c8865c8ee0d7d1c