[roll] Roll infra/recipes [tricium_analyze] Build all tools at once
Doing only a single ninja invocation will make it easier to convert this
recipe module to use "fint set" instead of "gn gen". The downside is
that we'll often build tools that we don't need.

However, we still won't run ninja if *none* of the affected files
require a non-prebuilt analyzer. Achieving this required some
refactoring in order to pass in a list of all files into the
tricium_analyze recipe module rather than passing them one at a time, so
we can immediately determine whether or not we need to do a build.

Original-Bug: 64654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/453413
Original-Revision: 876e7d56cf9273fa01fbde1ad2d38848bf64c74a
GitOrigin-RevId: e5c687a115dab4cffdacaeabdfd65556448dbf7c
Change-Id: I64f0e54a51c5c951c1ea4ec01552215f4be5616f