[roll] Roll fuchsia [fx][regen-goldens] Fix pattern logic and add tests
This CL fixes `fx regen-goldens foo` so that it regenerates goldens for
tools that have "foo" as a substring, rather than tools that are a
substring of "foo". It also adds tests using the Bash test framework to
catch issues like this.

Test: fx self-test subcommands/fx_regen_goldens_test
Original-Bug: 62471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/454494
Original-Revision: 5777ec1080114eba455f5fefc7564fe59c920798
GitOrigin-RevId: cc0e9f521e9917543887dd63d5bdf59be82bd0b9
Change-Id: I4c95792452439fc001bdf3295b6dd8b3b3538772