[roll] Roll cobalt Add jitter to Cobalt Config
The cobalt_app will pass in a jitter value between 0 and 1 representing
the amount of randomized offset added to the scheeduled interval. For
example, when jitter=.1 the initial interval (1 minute) will be
increased or decreased by a random value between 0 and  6 seconds. When
the current interval is close to the target (1 hour), the interval will
be increased or decreased by a random value between 0 and 6 minutes. If
jitter=0 the shipping_manager has no change in behavior.

On average uploads take place on the same schedule. But this will help
mitigate concurrent uploads and prevent leaking information about device
reboot time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/453703
Original-Revision: 92b92845a40634c762797ff6f5f11b93c285a3c2
GitOrigin-RevId: bcd86bc48f8b8f26082ea7a6ffc1000ea702bb5c
Change-Id: I9a69b8dc17627ed222c2c3914085ead42c819f87