[roll] Roll infra/recipes [coverage] Fail appropriately if profile builders fail.
If we fail to retrieve the orchestration_inputs_hash, that likely means
the subbuild of the profile builder failed. In this case we should
properly show that the underlying profile builder failed instead of
raising an infra failure for not being able to retrieve the hash.
Likewise, if there are no test tasks because the profile builder failed
to run and collect results for any of its test tasks, the covargs step
fails because there are no summary files. Instead, we should fail
earlier for not being able to retrieve any summary files from the
collected builds.

Original-Bug: 64985
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/396176
Original-Revision: 32352ffcea7d88c0f7db1fb594b32caf900b6aa6
GitOrigin-RevId: bdf9896ab2e244f2f05de88d3e949c393216af67
Change-Id: I1a207b98dd2777ca797c8f072587b562a923e5bd