[roll] Roll fuchsia [artifactory] add a flag -upload-host-tests
Currently the recipes don't build the host tests if they're not going
to be run, even if they're mentioned in tests.json. This can result in
failures to upload, such as seen in the attached bug.

This will be followed by recipe changes to set this flag to true
in the recipes whenever testing is enabled.

Original-Bug: 65244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455077
Original-Revision: 0aa32523fd63dbea10fbb2ddf99feb25b2e1ee81
GitOrigin-RevId: 6bade038864dc141f63700c9b104bdcdc20ac264
Change-Id: Ie25beebc370ed1fa830dcc7c20f6c3a2ebb356fe