[roll] Roll fuchsia [bt][host] Add Reason to disconnect requests, and log it
BrEdrConnectionManager::Disconnect() now takes an argument explaining
the reason for the disconnect request, and logs this as part of the
disconnect process. This aids debugging by easily indicating in the logs
why a disconnect of a peer happened.

TEST: This is a logging only change. To verify manually, disconnect a
peer connected over Br/Edr (such as via the `bt-cli` tool), and then
examine the logs with `fx log`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/454037
Original-Revision: 2da7bf9db0839be112434330351e204c243b7070
GitOrigin-RevId: 17073047a5f39f527861d94c5529a6849d66dfe1
Change-Id: I985a0cbe0d0360252bbe23aa244d2f3e1c561db6