[roll] Roll infra/recipes [tricium] Add support for "fint set"
Make all build configuration parameters properties instead of
hard-coding them in the recipe, and add a "static_spec_path" that will
be used to pass the path to a fint static spec for fint-enabled
builders.

This required some refactoring to run `gen` ahead of time at the very
start of the tricium analyses, rather than lazily running it the first
time we need access to GN results, because now we need to plumb in all
of the build configuration values from recipe properties, and this is
neater than the alternative of storing all the configuration values as
TriciumAnalyzeApi members.

Original-Bug: 64654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/453894
Original-Revision: ab24b57f584ae8c0ff43bde7714ff8a1eda2758a
GitOrigin-RevId: 9cc8e082697406e8db5d4609a7946bd000956631
Change-Id: I3597c0ed382930034b0a2984ee2b3ddd39e71a1d