[roll] Roll fuchsia [wlan] Limit the length of the partially hashed MAC
The length of the partially hashed MAC is limited so that the hashed
MAC is the same length of a real MAC. This reduces the length of
strings that log the hashed MAC to increase log readability.

Note that by the nature of the HMAC, truncating after some number of N
character bytes does not affect the properties we require to maintain
privacy with respect to MAC addresses. Additionally, by keeping 16
character bytes of the original 64 character byte hash result, we
ensure there is enough space to avoid collision with high
probability. In particular, there are 2^24 possible partial MAC
addresses and 2^64 possible truncated hashes. (Note in this case two
character bytes encode one actual byte of data.)

Test: Added testing of truncation itself
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452498
Original-Revision: 3ed5e63aea2a7affc683d4f4adab5e82d6fd05fb
GitOrigin-RevId: 3762e4c4f040b4e6ec19f3e88f7d62cd90c4e937
Change-Id: I6d2824c79a9c72781cc901e2494b7cfe0f2639a0