[roll] Roll infra/recipes [auto_roller] Use lowercase step names
Roll failure reasons should be lowercase and not end in a period when
they're presented in step names or step text.

For a slightly nicer UI, we'll still capitalize and add a period to
failure reasons that show up in Gerrit comments, for consistency with
comments added by CQ and Gerrit (e.g. "CQ is trying the patch." and
"Patchset 1 was rebased.").

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/455294
Original-Revision: ba9b1e33c99c8171e66a232da027fa6c0f45e5e3
GitOrigin-RevId: 780e9c21babe6931a4dfa63d4d70ed1be3abbcbc
Change-Id: I7d2f87286e77e38e9cca65e4b5c97fca760a715d