[roll] Roll infra/recipes [linux_sysroot] Support "fint set"
Make build configuration parameters properties instead of hard-coding
them in the recipe, and add a "static_spec_path" property in order to
pass a static spec to fint. I made it non-optional because there's only
one builder that uses this recipe and it runs very rarely, so I don't
see a need for a gradual rollout of fint here.

The only other change I made was to start setting `board`, as it was
previously unset (I think this was just an oversight on my part when I
first wrote this recipe as part of an on-call bug). That makes this
recipe more consistent with others, and will also make it easier to make
`board` required if we ever decide to do that.

Original-Bug: 64654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/455334
Original-Revision: b11c362e3bf71e1cb97b46dc47f27e47a3fd915b
GitOrigin-RevId: b1ff771ff2bffb516665b87e625c214c996b620a
Change-Id: Icdbf0204926fe633a2ceca49180b01eef8463974