[roll] Roll infra/recipes [kythe] Only do one `gn gen`
Instead of running a separate `gn gen` to produce the compdb after doing
a build, generate a compdb in the original `gn gen` before the build.
This will make the recipe faster and simpler, and make it easier to
convert to using the fint tool, which is much simpler to add to recipes
that only do one `gn gen` and `ninja` invocation.

Original-Bug: 64654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/455340
Original-Revision: c1856b3568b4f89020d21f477c29eb90629dc317
GitOrigin-RevId: af7a14ecd234c6f679e486e475b7cbd9fab0f83e
Change-Id: I5f859057d8cea98d174741dbd18cf05737a7ea17