[roll] Roll fuchsia [cpuperf] Increase test timeout
This test became very flaky on the afternoon of 2020-11-18, seemingly
because a lower-level change caused the cpuperf tool to become slightly
slower, causing the test to start timing out frequently.

For lack of a better option aside from disabling the test, double the
timeout from 1 to 2 minutes (the test is effectively unowned and we
haven't had luck with figuring out which change caused the slowdown).

Original-Bug: 63752
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455234
Original-Revision: 60cc30496d6df94cb5cc50852eb14cec4e906cf9
GitOrigin-RevId: df1f3a61ea4b4b1ead0e5aa3eac37142acb6793c
Change-Id: I2008b57fac5fb146982c091b625ab364ad0ec83e