[roll] Roll fuchsia [syslog] Add new vararg macros which take in file and line
The existing APIs will always produce the wrong file and line numbers
because of how varargs is implemented. The caller needs to provide both
of these details for them to be valid.

This is part one of a soft transition, necessary due to out of tree
users.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455162
Original-Revision: 83621b97b791926ffeef9ebe5d8b58e3718950a5
GitOrigin-RevId: d2c8e01119c15d792bb276428eb24cc84dca3961
Change-Id: I1bbb0593cf4b334138b0373431b5ff32501cb293