[roll] Roll infra/recipes [artifacts] set -upload-host-tests
This flag is added to aritfactory in
https://fuchsia-review.googlesource.com/c/fuchsia/+/455077.

Currently the recipes don't build the host tests if they're not going
to be run, even if they're mentioned in tests.json. This can result in
failures to upload, such as seen in the attached bug. It's only safe
to try to upload the host tests when the recipe has built them.

Original-Fixed: 65244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/454855
Original-Revision: 051c5c662446ce2f6866d012512dbcfd06d56516
GitOrigin-RevId: f5df68d383e67251d17ba39622c7a6bb588ae4fa
Change-Id: I98f90a7395d2fd62e08efbe002122bb175adefe1