[roll] Roll fuchsia [testing] Add method to look for bag of strings
Add a method to emulator which looks for a group of strings in
any appearance order. This is useful in situations where the test
writer wants to assert that multiple things happened, but the
order of their occurrence is not deterministic, perhaps because
they happen in different processes or threads.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455168
Original-Revision: 7f76485fce433bfe2f5f585310c729a41b278f63
GitOrigin-RevId: 404db899b2a407ab57485990872273da7ea857eb
Change-Id: I26715bb9a8f4b96d4a555fe322427bd097a03303