[roll] Roll fuchsia [docs][drivers] Mention the two sides of filtering
Explicitly call out the fact that there is both a producer and consumer
side to logging. This is a design decision of syslog, and some users are
unaware of it.

How to control the consumer-side of filtering should be elaborated upon
in the diagnostics logs, as it is general functionality of syslog.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455659
Original-Revision: 23f01108a0d24eb4fa0c6b0ca74fc9c6fced1a69
GitOrigin-RevId: f14b532766b4f4ab51d96d7a6a6eee65bb8b148c
Change-Id: Ie5a42f1465bc32980b36cb68096c06b070eb8261