[roll] Roll external/github.com/catapult-project/catapult Fix NaN that appears in RunningStatistics::meanlogs
When value == 0, std::log(std::abs(value)) = -inf. It is ok. Hovewer, when the second 0 is added we end up with (-inf) - (-inf) which is NaN. A histogram that is build in such way will fail to be uploaded, since the dashboard can't handle NaN in json (since it uses ijson library).

In Python code it is handled properly (the condition there is "value <= 0").

Original-Bug: chromium:1153044
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2560724
Original-Revision: 759827265102502cc7f814572675b9685c351908
GitOrigin-RevId: f20faaf9a34ed8d04c89e9d5c4593005e7d434e2
Change-Id: I74ed41826f0b3874e9d9de13e92f411cb370875a