[roll] Roll fuchsia [netstack] Fix endpoint add/delete race.
Add the endpoint to the endpoints map before registering for EventHUp
and starting loop{Read,Write} goroutines. The endpoint is deleted from
the map on an EventHUp. This change avoids a race where we get an
EventHUp notified and we try to delete the endpoint from the map before
the endpoint is added to the map.

Test: No new test added.
fx test -od netstack_bsdsocket_c_api_test -- \
--gtest_filter=LocalhostTest.RaceLocalPeerClose* --gtest_repeat=100000

Original-Fixed: 63864
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455916
Original-Revision: a9cb90d2814e879ec6f901abaaf580f7f0ce6283
GitOrigin-RevId: 1da90ab721bfb6a9c9dacc73ba54081488211a06
Change-Id: I68a69aca2c676991c85bb722b9383f06d7fe07ca