[roll] Roll fuchsia [mkzedboot] Avoid truncating regular files while writing to them.
While mkzedboot is intended to work on block devices, it mostly works
today on regular files too, allowing a zedboot image to be created as
a regular file as follows:

  touch zedboot.image
  truncate -s 16G zedboot.image  # Create a 16GiB empty file.
  fx mkzedboot -f zedboot.image  # Convert the file into a zedboot

This image can be used with emulators such as QEMU, or written out later
to a block device.

However, it only _mostly_ works today, because the "dd" command within
the script which writes out the ESP image will truncate the file,
destroying much of the image. (Block devices can't be truncated, so "dd"
won't truncate them.)

This CL adds the "conv=notrunc" parameter to "dd", requesting that the
output file not be truncated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/454675
Original-Revision: 950d6cc0a340a42376fa82d0aaf717f0195b0886
GitOrigin-RevId: 2dce905a698098c6d8e476cd738424ff0473faf5
Change-Id: Icec7f02df04c761d55b78bb4477b1d99248f3946