[roll] Roll fuchsia [bt][a2dp] Add ConnectedPeers::connected_stream
Add a stream to the ConnectedPeers struct which produces a peer whenever
one is connected. This allows for responses to peers being connected.

Use the new stream to add peers to the ControllerPool for AVDTP testing
when they are connected, and remove the plumbing used to previously
add the peers.

Original-Bug: 61530
Test: added bt-a2dp::connected_peers::test::connect_notifies_streams
Test: fx test bt-a2dp-unittests bt-a2dp-sink-tests bt-a2dp-source-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/454977
Original-Revision: a479f72415f009df2f9b7c7286b53382d9936efd
GitOrigin-RevId: 3ef4b3b4e1a415ead7761eb2a771fc84eefff6c9
Change-Id: I3b9afa8fa25dcca0671774f1dbcb88b8f14990ab