[roll] Roll fuchsia [a11y] Downgrade speaker error to warning
This error seems to occur when the screen reader "interrupts" itself and starts a new utterance while the old one is still being read. This is an expected situation and should not result in an error log. Long run we should change how this code works, but in the short term downgrading the log to a warning helps reduce error spam.

BUG: fxbug.dev/64419
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455554
Original-Revision: e6592855a423860d4c6fb1cffa6d7465152072e0
GitOrigin-RevId: 9029fe958db769557445575fbf1ef503810064e7
Change-Id: Ib7888f8340e8d54bf029c8fc9cac7559dbd8905f