[roll] Roll fuchsia [blobfs] Fix some hangs possible when encountering write errors.
When blobfs encounters write errors it was possible for it to hang. It's
not really something we need to support on real devices because there's
no graceful recovery if this happens, but it is an issue where we want
to do power-failure tests. I have tests that follow that exercise this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455902
Original-Revision: 5cddae4cfe68f835c5b92184b618826edcdd5283
GitOrigin-RevId: 2289b04dfd603637b369b47895ec9ba549f289ae
Change-Id: I7cc6153f0fde41678dbe80c86bbca4faef0ab0ab