[roll] Roll fuchsia [SetUI] Address flaky task manager test.
This fix addresses an issue comparing the beginning of
the longest running active task to the beginning of the
test. Due to the precision of the timestamps, it is
possible that the spawned task's start will not be after
the test start timestamp. However, it should not occur
before.

Original-Fixed: 65040
Test: fx test setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456474
Original-Revision: b058ae0a7deb7a2855d05c45644a1c2b8d0fc627
GitOrigin-RevId: 53a3b9cf23268bf7011b15fdef56539daf84c6a1
Change-Id: Ibf7583763bcbc36e1feafb79363f32c9fa475a5c