[roll] Roll fuchsia [web] Move sysmem to the list of required services
Previously sysmem was documented as required only for context instances
with VULKAN feature. That's not correct because video sysmem is also
used for fuchsia.camera3 API, which may be used when VULKAN is disabled.
The service is also likely to be required for other APIs and features in
the future. This change updates the API documentation to require sysmem
for all web.Context instances.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440263
Original-Revision: ed3e8e7180fe899636f5d64191caed1d60499760
GitOrigin-RevId: 692383cdd13963769e98db60ca4faeacd6af4767
Change-Id: I599431117b6d4eb4fc2f63b8b477c632d5f7d861