[roll] Roll fuchsia [fx][emu][gn sdk] Fix femu.sh failing to use KVM, fix internal `fx emu` dependency
femu.sh relies on `-a auto` which uses internal fx state - not present
in the femu context - in order to resolve this option for hardware
acceleration. Fall back to providing that bit of state ($HOST_CPU) if it
is not in the environment at the time of execution. Previously this was
causing KVM to never be used, and the emulator would run a lot slower.

Original-Bug: 62631
Original-Fixed: 64194
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453783
Original-Revision: 97535be2dc907c51e9a5ee14689d8c1f7be32d93
GitOrigin-RevId: 4b07295177c6578fdc74ce94653ab04bf4a27de8
Change-Id: I96f71605dfbc1b76cdd8ed039115845c6d43b58b