[roll] Roll fuchsia [component_manager] Introduce CapabilityTypeName.
Introduce a new type to cm_rust CapabilityTypeName that provides a
bounded enum for the previous &'static string that was returned from
CapabilitySource.type_name(). By converting this type to an enum it can
be more efficiently compared by the PolicyChecker in the hot-path of
capability routing without incurring the cost of a full string compare.
It also prevents having to deal with the unbounded possible values of a
string when the type_name is infact bounded to several types.

Original-Bug: 61866
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453824
Original-Revision: 3a6d11011b05d6e1337dc06858d3e93dfb325e4c
GitOrigin-RevId: a77b77699f90799ef0053f0e4962c6ea9de0f6a9
Change-Id: Ie14f7281486b63a769a9b669d1c9eea37e31aeb4