[roll] Roll fuchsia [virtualization] Move configuration to runners
Modify the VMM code so that configuration is handled through runners.
This means that files are opened and passed through LaunchInfo from
runners to the VMM, and the VMM is unaware of command-line options or
configuration files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456280
Original-Revision: 5ad7c54a8238f5c64c0a5d6e6113c43672a93d98
GitOrigin-RevId: deae131f134906dc0c10e971a321844868ba3534
Change-Id: I24446f466899cf5dd19aae9807873f1b27b36b77