[roll] Roll infra/recipes [goma] Fix "arbritrary" typo
This is potentially a breaking change because it renames a property.  I
checked Code Search as well as the Flutter configs and couldn't find any
usages of this property, so I think it's safe to rename it to correct
the typo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/456341
Original-Revision: 8c6bc1c4d2fb6cdaf79a5260f9a6709faff943fc
GitOrigin-RevId: b173018d69793115902753da2d9d0d0c235302d8
Change-Id: I39174269c6c30fd6542111226ce87e79cc6af424