[roll] Roll fuchsia [fint] Delete collect_metrics static spec field
Metrics collection is now enabled unconditionally. This won't be safe to
land until we stop setting this field for all builders (since static
specs are stored in textproto format, it's safe to renumber fields but
not to remove fields that are still set in some textproto files).

Original-Bug: 64654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455877
Original-Revision: 524093a8669069dfee7812655ec94930dd2686bb
GitOrigin-RevId: 9655fbefa2959a597c5ff9a0a56fba9f8e0edbf5
Change-Id: I7c7a6cd91d0feec55204ffff9db4aac2d6dbc237