[roll] Roll fuchsia [blobfs] Integrate the external decompressor.
Integrates the sandboxed decompressor and adds options for enabling it
which default to off for now. Adds metrics to count how often the
decompressor is used to the read stats.

This makes a change to greatly simplify the SeekableRemoteDecompressor
as the previous interface could not integrate properly given the way
that the pager places data in the vmos.

Original-Bug: 56464,62177
MULTIPLY: external-decompressor-runner (fuchsia): 20

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455355
Original-Revision: 28f38048d62af34297f092911687c1dca448b75a
GitOrigin-RevId: 1b8026acf6b937a31de01a3a748107421408d384
Change-Id: I5d604829967e61d47d7116eb6ed030cc69dff80a