[roll] Roll fuchsia [artifactory] Upload GSYM to GCS
GSYM is a more efficient format for debug symbols that we'd like to
start using. In order to do that we need to upload GSYM debug symbols in
artifactory, to both the legacy "debug" and the new "buildid"
namespaces.

This will be a no-op until we start setting an `output_gsym` arg on some
builds and populating the "gsym" field in binaries.json to point to each
binary's GSYM file.

Original-Fixed: 61677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456547
Original-Revision: a95724c479bd0fbcd42a7d4e64d55f388c1b59d6
GitOrigin-RevId: 3ce45fcf46b5ad427d80a72aaecd59c21d8d2835
Change-Id: I9b91e37ce7c771365c7731964cfdb73aac7f3d9e