[roll] Roll infra/recipes [tree_closer] Include full build link in bug text
Use the fully qualified build URL in auto-closer bugs so that it's
easier to see at a glance which builder is failing, and for consistency
with links in other places.

Also link to the builder page of the last 200 builds in bugs that report
consecutive failures by a single builder, to reduce the number of clicks
required to vewi the builder history. I didn't add this history link for
bugs reporting concurrently failing builders because we'd need an extra
link for each builder, which would clutter the bug description. I might
eventually add it if and when Monorail supports markdown natively, which
would make it much easier to do things like this while keeping the bug
text readable.

Original-Fixed: 64505
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/456767
Original-Revision: 0755ffc0941d6a43a358fd88156981dca1686232
GitOrigin-RevId: 41ace01934a546da0ba67a255fa85ebeced5b2dd
Change-Id: Ida2b1ecd97afbbc18f9624fc1cc068fb25dea140