[roll] Roll fuchsia [kernel][mexec] Avoid allocation overlap
Passes the end of the target kernel image location to the identity
allocator to avoid overlapping allocations. This is neccessary since
the destination for the kernel image is not selected by the pmm and so
it otherwise does not know how to avoid it.

Without this the included asserts can trip, and without the asserts
this overlap can result in mexec failure, typically manifesting as a
triple fault on x86.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455901
Original-Revision: ad4daf0e7cb02dea79e398d81dac051d759b147f
GitOrigin-RevId: f0f1b5849b289626f19379f5c7e0cea5466aed8e
Change-Id: I0cf70a1f33c99f6d91b525ddbf5d84ae7f5aaaaf