[roll] Roll fuchsia [musl] Don't use empty initializers when building for C
In C, it is technically illegal to initialize a struct by saying
something like

  MyStruct my_struct = {};

This said, GCC will allow this by default, as will Clang. If, however,
someone is building using Clang, and -Wgnu-empty-initializer is
passed, this code will generate a warning.

Currently, the definitions of some of the pthreads structure
initializers in our version of MUSL make use of this empty-initializer
extension, meaning that 3rd party code might run afoul of this
technically illegal pattern, depending on their build settings.

So, avoid this issue by initializing with { 0 } instead of {}.  Note
that this will not work in C++, nor is using designated initializers
an option as those are not officially supported by C++ until C++20
(and we currently only require C++14).  To work around this, two
versions of the initializer are provided, using the __cplusplus
preprocessor macro to choose which to use.

Test: new unit test
Original-Fixed: 64794
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456734
Original-Revision: 1d8e90376fb351b9427aae838ccde713a22e1b44
GitOrigin-RevId: a7f946c95ed28446df4b3f71f673ad788f584ccb
Change-Id: I577b1f7ac6c19b3cff784f16b4af5b43116897c0