[roll] Roll fuchsia [fdf] Clean up ZIRCON_DRIVER macro style
Remove the semicolon from the end of the macro to get the compiler to
enforce writing one at the end of the callsites. This also prevents the
compiler from complaining about empty statements in certain cases when
there already was a semicolon at the callsite.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457074
Original-Revision: 3479ca20e489bfcfc6491a05612ab53cfd11e7f2
GitOrigin-RevId: bee32e456e201942e3e6c405873b5cd9c65ab4e1
Change-Id: Id16bc5f84d8541bb79b308647f929c368dc98364