[roll] Roll fuchsia [storage] Update fs_management rust library
Build on fs_management and add support for per-filesystem
custom arguments.

The fs_management library cannot easily be updated because
it is used in `vendor` portions of the fuchsia tree.

As a result, `Filesystem<Minfs>::set_readonly` needs to exist
for compatibility reasons. Once this CL lands, the vendor
code can be changed and this function can be removed.

Test: fx test fs-management-rust-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/454975
Original-Revision: ac664342072ce2715dc1c6e360edc899d8c75f8d
GitOrigin-RevId: 78624607a995e412f727070a26d38bd58fb55120
Change-Id: I6b9391b877ef4c17d6d89959bcab4dd136bb9ef4