[roll] Roll fuchsia [fidl][llcpp] Rename the internal EventSender to WeakEventSender
The next CL introduces an EventSender class that actually owns the
underlying channel, hence avoids the unsafe borrowing that is used by
the static event sending methods.

Since the current EventSender is actually a weak_ptr to the binding
object which owns the channel, I'm intending to rename it to
WeakEventSender to clarify.

Original-Bug: 65212

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455220
Original-Revision: 44a78a30676109845bcadda0b919a146cf4c331a
GitOrigin-RevId: 7caa686b20188d0a4b12221112c823477e3d2657
Change-Id: I3092528ff9c87ff886e7cfe85f9e06162454558d