[roll] Roll fuchsia [hlcpp] Split Message class into HLCPPIncomingMessage and HLCPPOutgoingMessage
This is part of work on FTP-028. The classes need to be split because
we are now using different handle types (zx_handle_disposition_t and
zx_handle_info_t) in the incoming and outgoing directions.

The naming "HLCPPIncomingMessage" and "HLCPPOutgoingMessage" is based
on the equivalent "IncomingMessage" and "OutgoingMessage" used by LLCPP.

This is part of an LSC. In this CL, parts of "Message" are retained
for use by out-of-tree users. In subsequent CLs these usages will be
migrated to the appropriate "HLCPPIncomingMessage" or
"HLCPPOutgoingMessage" object.

Original-Bug: 41920
Original-Bug: 65523

Testing: existing tests migrated, lateral change coverage-wise

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455615
Original-Revision: d083fc18d72e1151188ac2b1b10265efc0c8906b
GitOrigin-RevId: 15a47f3205636470674de733801a61acd72956b4
Change-Id: I688ef9556fc4548f9a4aa0fd30dde063230169bf