[roll] Roll fuchsia [fidl][llcpp] Introduce MyProtocol::EventSender which owns the channel
Since MyProtocol::EventSender owns the channel, users can safely send
events through this object, without worrying about the channel becoming
destroyed in the process.

All uses of the static SendFooEvent functions should be migrated to
either use this if the channel is not bound to a server,
or use the fidl::ServerBindingRef if the channel is bound to a server.

TEST: Added one testing sending events using this object

Original-Bug: 65212

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455221
Original-Revision: 173b8e3ace2d000a55f27b40d55cbbb618b395dd
GitOrigin-RevId: ed41eb63ccb09d203b1ab18ba64a467ab99add05
Change-Id: I77caa83164639f5aeca1c8cc7d759683faa94abb