[roll] Roll infra/recipes [fuchsia] Delete api.fuchsia.test() helper
Most of the logic in this test helper function involved setting default
fuchsia.py recipe properties that are ignored by most recipes that use
the helper function, including the fuchsia.py recipe itself, now that
most properties are passed via specs.

The function also accepted a list of step data as an alternative to
adding step data to the test object, which is inconsistent with how
other recipe tests normally construct test data, so I think it's an
improvement to stop using it for that purpose.

The only feature it provided that is still useful is a `tryjob`
parameter to control whether the mock Buildbucket input corresponds to a
CI build or a CQ build. But this isn't even required by most recipes,
since most recipes don't significantly vary their behavior between CI
and CQ. So we can create small recipe-specific helper functions for the
few recipes that do use this feature.

The expectation file diffs are entirely due to removing old unused
properties from the mock data.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/457154
Original-Revision: b498f7edd56efdc3f7bac3cd8a5e2398196ba064
GitOrigin-RevId: 52026a82e7b9e275c9d351b81840d9be52fb5671
Change-Id: I51b0f41a66d4c7de9ed2aeb618784e681ad2d6c6