[roll] Roll fuchsia [dma-buffer-test] Use |options| parameter in zx_vmo_create
In the refactor to make dma-buffer-test use lib/fake-object a
mistake was made that led to the options field being ignored
in the proxied zx_vmo_create. Since the only option available
to that syscall is for resizing it wasn't a problem in the tests,
but this fixes it.

Original-Bug: 53888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457236
Original-Revision: 045e2dcbada160104b56ce9e239c230befdf6103
GitOrigin-RevId: 440133d61331bf28a7fa3a9c143965ccfcaa16b4
Change-Id: I0e83ee993df174e916cec01da09e6bdbc236a218