[roll] Roll fuchsia [zbitl] Cosmetically update fuzzers
This change updates the iteration fuzzer to have a simpler and more
idiomatically-spelled control flow - and updates the names of the fuzzer
to include a -fuzzer suffix. The reason for the suffix is that host-side
tests are created automatically by the fuzzing framework that could
otherwise pollute the host artifact build directory namespace with names
like "iteration-test".

Original-Bug: 62638
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457265
Original-Revision: 6a298e70cf6efe083a67f0f84524ad9c0ea89aa5
GitOrigin-RevId: 9353149e82818c73d4d28f0c5d91bd3ff83084cd
Change-Id: I1592565893deb60dd037f5f8aeb1c89a359b5ff5