[roll] Roll fuchsia [audio_core][tests] Pipeline fidelity tests, pt.2
This is the second part of a refactor to the hermetic
audio testing library, to add fidelity testing. This
needed to be landed in stages, because this lib is used in
other repositories. With this CL, the refactor is complete.

This CL also contains additional renaming not done in the
initial CL.

Test: test-only code, no product change
Multiply: audio-core-google-pipeline-tests-astro
Multiply: audio-core-google-pipeline-tests-sherlock

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453183
Original-Revision: 3a729112b5f654941160626fa0fce2e11994e43e
GitOrigin-RevId: 66a4cd5f38b0487992f53118b696b6486d117d25
Change-Id: I69c9ce03ff59e0e347a024a01ac79ae000289577