[roll] Roll infra/recipes [build] Propagate gn gen failure summaries from fint
fxrev.dev/457238 adds a new `-failure-summary` flag to fint, which
should point to a file to which fint will dump the stdout of `gn gen`.
For consistency with the `gn gen` recipe code path, we should propagate
this output to the build summary markdown in case of failure.

I also renamed the "fint set" step to "gn gen" to avoid confusion when
we start deploying fint more widely. I can imagine folks might get
confused if they go looking for the "gn gen" step but can't find it.

A better long-term solution for this might be to have fint implement the
luciexe protocol so that its subprocesses show up in the build results
page with their own step names. But this probably isn't worth doing as
long as each fint command only runs a single subprocess.

Original-Bug: 65704
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/456770
Original-Revision: 999378eca4b8245047492336d202c9ce6dff73e4
GitOrigin-RevId: 5c6d1ab359b33f314e25b5e153f71f83a5714bda
Change-Id: I1df1d80149335762c36375697707cdcd2927ebcc