[roll] Roll fuchsia [fatfs] Check for overflow when calculating sector number.
Fatfs incorrectly assumed that multiplying an 8-bit number by a 32-bit
number would never overflow the multiplication. In a well-formatted
filesystem this would be true as the number of sectors is a 32-bit
number, but the fuzzer uncovered this issue with a poorly formatted
filesystem.

Original-Bug: 65490
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457456
Original-Revision: 4efdce10c89f55e67050cc44a54837c2a087391c
GitOrigin-RevId: bb2a74c7c63e1de7541c17507d2b00041a6adadb
Change-Id: I33a856744f0724dc68dbb2126dd91fdbb71af17a