[roll] Roll infra/recipes [ffmpeg] Support `fint set`
This is the only recipe that does two builds with differing build
parameters, so to support fint we need to pass it two fint params files,
on for each target architecture. During the transition period we also
need to pass all build configuration parameters as properties, so I
created a `Build` property message that encapsulates all the values
required to do a single arch-specific ffmpeg build.

Unfortunately, we'll always need to pass `target` and `variants` as
properties even after fint is complete, because the recipes uses those
values to locate build artifacts to include in the ffmpeg CIPD package.
Keeping those properties in sync with the fint params file indefinitely
will be slightly awkward, but not super hard since this recipe and its
corresponding builder are rarely touched anyway.

Also make the `manifest` and `remote` values properties, for consistency
with other recipes.

Original-Bug: 64654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/456771
Original-Revision: 3f78ed99fe9085de6e6ab3591842998bf9402e8f
GitOrigin-RevId: 352490ebc03e3ffaf96345c809f0e7d819dcf724
Change-Id: I6f74a5cc431e35fcca14c42d5ff4bd7f76d60863