[roll] Roll fuchsia [fidl] Add source compatibility scaffolding tool
This adds a scaffolding tool to help generate self-contained
source compatibility tests - see gen.main.py for usage.

An example output from this script is added under
protocol-method-add. One file of interest is the test.json
file, which is a serialized representation of the test
inside the directory. This file can then be used by other
tools, like the source compatibility test GN template for
building the test or a documentation generator (both are
being added by ianloic@).

This tool is mostly still WIP - the next steps are to
land the GN template, write/run some tests, and use that
to iterate on the tool.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441055
Original-Revision: 9c743a8698652fe17c8591a055ace57901814d1d
GitOrigin-RevId: 5b6349b0d003af6aead5050cdb8c97585fcc8fb0
Change-Id: I5d80963d24f56a03c4c38348aac3322e2caa9f05