[roll] Roll fuchsia [cm_fidl_analyzer] Introduce the cm_fidl_analyzer library.
This initial CL just integrates the new library. Currently the library
has no analyzers but adds the library to the tree. Some examples of
analyzers we want to support long term:

Full tree topology verification. Given a set of all component manifests
verify that each component in the topology has correct use/expose/offer
statements. For example you are not using something your not offered (in
the context of the provided topology set).

Individual topology verification. Geven a set of component manifests
verify that your component is valid in what it offers,uses,exposes.

Additional linters, prefer this protocol over that. Upgrade to this
protocol from that etc. This will ease the burden of migrating things as
we may be able to directly integrate these locally scoped linters
directly into CMC as an optional step.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457280
Original-Revision: 340783fe05358b55eca1e3624353fb132923cdce
GitOrigin-RevId: d7fd54c8e7e9b8ff16d22ee1dd0a6de19a7ff100
Change-Id: I1e2d58986344a55699b0eb9916b480f28ec3d2a7