[roll] Roll fuchsia [e2e][intl] Increases wait between retries from 1 to 3s
Increases wait time between probes for timezone change in the timezone
test from 1s to 3s.

This change is based on an assumption that flakes due to long message
latency are simply long latency tails, and not a more complicated issue.
The change is a harmless way to disprove the long latency assumption: if
the flake remains, we need to dig deeper, but if it is resolved, then
we got a quick and easy fix.

Original-Bug: 65318
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457798
Original-Revision: b05febf1c77d5060f2f36b6a1cc7ecf9764beb1e
GitOrigin-RevId: 9311b3632895c22bc8bde22b34199e35b21b6684
Change-Id: Id1476f8f2b867cee756344c8d724475374169550