[roll] Roll fuchsia [gigaboot] Raise initial EFI allocation from 6 MB -> 8 MB
gigaboot requests a fixed allocation for the kernel load
zone; the allocation size was a constant, which was large
enough for plain kernels but no longer large enough for
kASAN kernels when global instrumentation was added
(6.5 MB vs 6 MB).

Raise the initial allocation request to 8 MB to cover
this for now.

Fixes boot of kASAN + global instrumentation kernel on NUC.

Tested:
* Booted --variant kasan kernel + zedboot on Intel NUC6i5
* Booted plain variant on GCE; a --variant kasan kernel
  will not work on GCE, either before this CL or after.
  GCE can only support a <= 2.5 MB EFI allocation at our
  load address.

  GCE continues to work because of a fallback in the code;
  if the initial allocation fails, we step down to a
  smaller size.

A real fix would be to make the kernel relocatable or at
least parse the kernel to find the appropriate zone size
and request that. This is just a targeted fix to make the
kasan trybots work again.

Original-Bug: 30033 Kernelmode ASAN for Zircon
Original-Bug: 58770 KASAN Fails to boot on NUC
Original-Bug: 32374 Make x64 kernel relocatable
Original-Bug: 32223: Gigaboot's pre-allocation of the kernel range gets
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421177
Original-Revision: 8c5145a07768c98a58bd0268638089d41901ecc0
GitOrigin-RevId: bdb47c1bf2dbfb401072259647f632d6f16f9de4
Change-Id: Idde377370e0d97c9fab1f9a6e28d15e95b9ea3f7