[roll] Roll fuchsia [SetUI][Audio] Use flag to signal change.
The Audio controller currently uses a timestamp to signify
a stream's volume has changed. This logic depends on
the precision of the timestamps being greater than the
time between requests. This turns out not to be true in
some environments, such as tests. This change converts
the signal to a flag, where an alternation between high
and low indicates a change in the stream's volume.

Original-Fixed: 65747
Test: fx test -o setui_service_tests
Multiply: setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457672
Original-Revision: d73e104b92eae1a83f75a85891acc6214cafa33d
GitOrigin-RevId: 49e65eefc49bef74e72174e1e7780347ad4572d9
Change-Id: I5e75dae6faf84085b55d3ace3752fae4a55529e4