[roll] Roll fuchsia [zbitl] Add a fuzzer for copy codepaths
Several bugs were found around bounds-checking and with the
copy-creation byte range method (specifically, with the
creation-CopyWithSlop method failing to return the proper slop value in
the non-clone case and failing to zero out `to_offset`-many bytes in
the created storage). To address the latter, we updated the behaviour
of the Create() trait to do the zeroing out of the head of the storage.

Fuzz support for the storage copy codepaths will be added in a
follow-up.

Original-Bug: 62638
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456117
Original-Revision: fccf0b422b3c13877fa1e70a77f656853856b442
GitOrigin-RevId: 620ae04e5c470bb3526b2d0e7157b8daabc62eac
Change-Id: Ie78b01bdb4591f5df92e324cb72536752db161d8