[roll] Roll external/github.com/catapult-project/catapult Implement the unified V8 and CppGC metric computation
See bit.ly/v8-gc-stats-collection for the background and motivation.
The central part of the implementation are rules that specify which
events contribute to which metrics (see RULES in the code). This
decouples the metric computation code from event names and event layout.

New metrics can be added by adjusting the METRICS and RULES lists
without touching the metric computation code.

The new metrics are not enabled yet because we need to first land a CL
on the V8 side that adds the epoch argument to V8 events.

Original-Bug: chromium:1154636

Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2563467
Original-Revision: 88efddc6a162994babc45a2d28bb5f4f9867c2f1
GitOrigin-RevId: 055aa577629c001d46594ea6a80910f274f6fb92
Change-Id: Ifcdfd05adafd0ff2ff6f1403bebb5407bbf942be