[roll] Roll infra/recipes [roll_comment] Skip changes without Reviewed-on footer
Fuchsia-owned Gerrit hosts add a Reviewed-on footer to the commit
message of every change that goes through Gerrit, so if a change is
missing a Reviewed-on footer then either it came from another Gerrit
host, or it didn't go through Gerrit at all (e.g. a merge commit). In
either case we shouldn't try to comment on the CL after it's rolled,
since commenting will likely fail due to insufficient permissions on
other Gerrit hosts or the change not existing in Gerrit at all.

Requiring that the footer be set also lets us skip the `change-details`
step to retrieve the change number. Now we can get the change number
just by parsing the Reviewed-on URL.

This change should solve the failure seen in this build, which resulted
from trying to fetch change details for a merge commit in the Fuchsia
fork of Go: https://ci.chromium.org/b/8861846506975441184

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/458194
Original-Revision: 26ea1a4fe64fbd219c8ac6076dfeca0263d04e56
GitOrigin-RevId: bcafaeed8ef51ebeee871744f943dfc3577686b9
Change-Id: I03ebc4beb284c825f5fb73b330a2d1ee3dbbe4f5