[roll] Roll fuchsia [wlan][sme] Handle bss_desc in scan and connect.
Populates the bss_desc field in scan results, and updates SME
to bypass the join scan and connect directly when passed a
bss_desc in a connect request.

This is the final CL to support connection by BSSID in core. We
will revisit this logic if/when we make connection by BSSID the
only supported emthod.

Original-Bug: 65481
Test: Updated/new unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452513
Original-Revision: 3e84922c47b5578e52e62f77b2741564de0ab749
GitOrigin-RevId: 1ecabd743f0b37c93032c81f6f7a026b3d4406ce
Change-Id: Ifc53ba6153d45a1ebff5ca59388cd1a8251ae01a