[roll] Roll fuchsia Reland "[component_manager] Actions use SharedFuture"
This is a reland of a811c6a1055eff0b58489548924ebe701936d190

Difference from the original change: Instead of sharing the future for
the action, the action future is spawned in a separate task, which sends
a message over a channel. The receiving end of this channel is wrapped
in a shared future, which is returned from `ActionSet::register`. The
bug was that we could not guarantee the shared future would always get
polled -- they could be aborted, for example. Not only did this mean the
action did not always complete, it meant that sometimes the shared
future was left holding a mutex, causing deadlocks.

MULTIPLY: test_manager_test: 300

Original change's description:
> [component_manager] Actions use SharedFuture
>
> Actions represent an asynchronous operation on a component or realm
> (such as stop, start, and delete). Actions are idempotent, in the sense
> that two matching actions scheduled at the same time will only be
> executed once.
>
> Currently, this is implemented as a custom future. Change the
> implementation to use SharedFuture, which is simpler.
>
> A side effect of this change resulted in task ordering differences. This
> caused some tests to fail which were inadvertently dependending on the
> previous order (for example, assuming that when bind() returns, the
> runner has actually started the component). Fix those tests to be
> agnostic to the order.
>
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450182

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451957
Original-Revision: 691da1994e39ccea16dfd95372234ac9dafc1c88
GitOrigin-RevId: c21a09b73cc9e2baae2975a2ddbb172560ab9fdf
Change-Id: I3c4a517702cd4dd98be6f39495e49f47854eb379