[roll] Roll fuchsia [kernel][sched] Fix an edge case in changing to a deadline proifle.
An edge case was introduced when adding support for heterogenous
processor performance. In that change, deadline tasks are given
proportionally more time when running on slower processors to
ensure they get the same effective amount of work done before the
deadline. Because scale factors were added to various calculations,
a debug assert was added the common reschedule path to ensure that
reaching the target preemption time is consistent with the scaled
time slice expiring.

When changing from a fair profile to a deadline profile, the
remaining time slice is clamped to the capacity of the profile to
prevent violating bandwidth isolation, while still permitting the
thread to run when there is time remaining. If the thread is
running at the time of the profile change (e.g. it is setting its
own profile), then the target preemption time computed from the
unscaled time slice when it was a fair thread becomes stale and
will assert when compared to the scaled time slice, due to being
a deadline thread now.

Fixup the target preemption time to account for the performance
scale of the CPU it is running on when switching to a deadline
profile.

Original-Bug: 61508
Test: Looping fair <-> deadline on a little core does not assert.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458020
Original-Revision: edf452415f2f62d17bc31788a561c7109bf8cebd
GitOrigin-RevId: 1c125497071ef8bf447275abb682ab0dadd4811b
Change-Id: Ic055fd69c23ac07c61407d669e8f3a9ed3a56c7f