[roll] Roll fuchsia [fdio] Implement ftruncate for memfd
This implementation is somewhat racy because we need to touch a few
different aspects of the VMO to resize it accurately. We have a similar
problem in memfs. It's likely we'll need a new kernel API for doing this
work atomically, but we can refactor this code to use that API when its
available.

Original-Bug: 60236
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457948
Original-Revision: 0acab74485693d12d842a9a0aad20470577feb39
GitOrigin-RevId: 7ac6679fcefd993cd93d8d2cce74288e48b015b3
Change-Id: Ibc447cc7e6f10e5b1baa40c1338fcd960d937e09