[roll] Roll fuchsia [blobfs] Support page requests larger than 256MB
Add support to page in requests for ranges larger than the transfer
buffer and the decompression buffer. zx_vmo_op_range with
ZX_VMO_OP_COMMIT on the entire blob generates a page request spanning
the whole blob. The transfer and decompression buffers are sized at
256MB, which limits the size of blobs that can be committed entirely to
256MB. This change removes that limitation. The user pager now reads in
and decompresses the range in chunks, so that they can be accommodated
in the buffers.

In practice this code path is not expected to be exercised often, since
it only applies to blobs larger than 256MB.

Test: blobfs-unittest

Original-Bug: 64454
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456735
Original-Revision: a78211c1b7c6348bd69514b1586c47cdd172ab39
GitOrigin-RevId: de27b072fdb086e5af3fe36f48cbd19a73b793dd
Change-Id: Ie77a02b34d55b00d1bb79fc62a91bd4b5b609419