[roll] Roll fuchsia [mac][vulkan][build] Fix libvulkan.dylib symlinking
Previously the dylibs were implemented as:
1. hardlink prebuilt/.../libvulkan.1.dylib -> out/.../libvulkan.1.dylib
2. softlink out/.../libvulkan.dylib to libvulkan.1.dylib

This didn't work with gn stamp files because step 2 touches a stamp
file, but when the timestamp of its input is looked up, the link is
followed back to libvulkan.1.dylib, and that in turn shares a hardlink
mtime with the dylib in prebuilt (which isn't modified so depends on
when cipd synced it.)

As a result, the mtime of the input libvulkan.dylib is always "older"
than the .stamp file, so the build is always dirty.

To resolve this, change to a plain copy from prebuilt -> out, and in the
same step, generate the symlink. Remove the symlink.py script and make
it a custom helper for this purpose as it's (as evidenced here) a bit of
a footgun in GN/ninja.

Original-Bug: 65872
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458287
Original-Revision: 4be6ade9a265a1e5d38044675b64a9ff0d54d351
GitOrigin-RevId: 7506b7b4d6cfaafc9b8425d6a4bcbe0a191f5efc
Change-Id: Ic22ae1f309b9ebb4752726021d263578bd6b9446