[roll] Roll fuchsia [bt][bt-gap] Errors restoring host bonds do not fail Commit()
When processing the fuchsia.bluetooth.sys.bootstrap.Commit() function,
do not return an error if storing the data in the store succeeds but
restoring the data to an active host fails. Instead, this is treated as
a Host error but not an error with the commit procedure.

In order to test behavior correctly, also implement an in-memory version
of the bt-gap Store that doesn't use the Stash component or protocol,
and so can easily be unit-tested. Migrate existing tests to this
implementation and remove the `stub()` function that previously existed.

Original-Bug: 63013

Test: fx test bt-gap-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/455597
Original-Revision: d27690d643166178c8861d7b2497118a5a8d7907
GitOrigin-RevId: 997aaf63712b2a618f9e40e8eda9421d251bc992
Change-Id: Ic380368b2c76e405dc148552b9d31e1836c970c6