[roll] Roll fuchsia [doc-checker] restrict yaml checking to _toc.yaml
RFC publishing will use yaml files to store metadata, and they are
not part of, nor compliant with fuchsia.dev publishing yaml schemas,
but the current doc-checker code verifies any _*.yaml file.

This CL changes doc-checker to perform fuchsia.dev related tests on
_toc.yaml files only, and just perform standard YAML syntax validation
on YAML files that are not named _toc.yaml.
Although it is technically possible to have fuchsia.dev TOCs with
different names, by convention we only name them _toc.yaml,
and other pieces of the publishing infrastructure rely on them
named _toc.yaml, so it is safe to assume that the name inicates
whether a YAML file is part of the publishing pipeline or not.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458279
Original-Revision: b247065d95f4e9fb372b86ff2d7fd37799b1e735
GitOrigin-RevId: d266f47b487c7e4e5714fb9c7b8c6b2d65c2c59d
Change-Id: I1d81bea3fe8db938a01293261879ac211af46daa