[roll] Roll fuchsia [hlcpp] Remove Message class
This is part of FTP-028 and an LSC to migrate from the Message
class to HLCPPIncomingMessage and HLCPPOutgoingMessage.

This is the last step of this LSC -- to remove the Message
class. "fx lsc presubmit" has been run on this CL and it should
not cause breakages.

Original-Bug: 41920
Original-Fixed: 65523

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457616
Original-Revision: 56f43a4c051d406ee7b26773bb246710516acd6a
GitOrigin-RevId: 43bccf11964897f6fd0fbce02dde61a232f73015
Change-Id: Ie9226040b7ce2aadb4b30161cd6d035c81a8055a