[roll] Roll fuchsia [fidlc] Fix crash on nullable bit fields.
Bits should behave identically to enums when used as nullable fields
in structs; that is, the compiler should error and inform the user that
this behavior is banned.

This reporting was not happening for bits, and was resulting in a crash
instead, due to a missing case in the switch statement that validates
that a token is nullable.  This case, along with a test for this bug,
are added in this changelist.

Original-Fixed: 40994
Test: fx test fidl-compiler; fx test fidlc_golden_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458345
Original-Revision: e9e1af75bbc2db0a2c0554cae381fe566f343261
GitOrigin-RevId: 5902785901a7b8f41ab3d7a5067af66fd98898a3
Change-Id: I83756afbdb519847fdae1d9f70e0723fd4b169ad