[roll] Roll fuchsia [firmware] flesh out README
Adds more detail to the README, such as details of each included
component and general porting instructions.

We will be adding this to the Firmware SDK shortly, so this will be the
main reference for developers to port it to new bootloaders.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458336
Original-Revision: fcbd012813f484c1a1a48684aeb8c7d1cf76cd18
GitOrigin-RevId: 608ca5316cc0894523943927f3d53c319ef08332
Change-Id: I950f0b78ea33cee237f8a80ba7cb99272fb07a61