[roll] Roll infra/recipes [auto_roller] Retry `git push` steps
Rollers sometimes fail to push changes after restoring previous roll CLs
due to replication delays between Gerrit backends: the backend that
handles the push might not yet know that the change has been restored.
For example: https://ci.chromium.org/b/8861661267313107008.

Retrying push steps (up to 5 times) with an exponential backoff should
generally fix this by allowing more time for the updated change state to
propagate between Gerrit backends.

This required a change to the behavior of the `api.utils.retry()` method
to have it return whatever the wrapped function returns, because we need
to be able to add a link to the push step. Previously the `retry`
function required that the wrapped function return a boolean indicating
whether the operation was successful, which is a less intuitive API in
my opinion. The auto_roller recipe module is currently the only user of
that function, so it was an easy change to make.

Original-Bug: 65966
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/458715
Original-Revision: 808acf2e7fd9f069cebcc46df543e8c73103ad0e
GitOrigin-RevId: 7fee1ac91dd61b1cfa9fa766f8f4f1216509baf7
Change-Id: I0aa946fd131b70ee5c7690f03c57b23ecdc40fc1