[roll] Roll fuchsia [netstack] Do not store endpoint with hard error
Do not add an endpoint with a hard error state to the endpoints map.
The endppoint is removed from the map only on an EventHUp, and such
endpoints would no longer see an EventHUp.
For ex: Accept of an endpoint that was reset post established state.

Original-Fixed: 66033
Test: fx test -v netstack_test -- -test.run=TestTCPEndpointMapAcceptAfterReset

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459335
Original-Revision: 86d9317502df3b3dba15cd189f3fa021107eb215
GitOrigin-RevId: 0337988b03bc9c683f47971ca98d98e298740ac5
Change-Id: Ibc5e729d36f8d4defe8bab339e2331c9575c62d0